# perf(detection): Improve file read logic