--- a/pkg/detection/detect_test.go
+++ b/pkg/detection/detect_test.go
@@ -187,6 +187,52 @@ spec:
 			},
 		},
 		{
+			name: "kubernetes, reader, JSON",
+			path: "k8s.json",
+			r: strings.NewReader(`{
+  "apiVersion": "apps/v1",
+  "kind": "Deployment",
+  "metadata": {
+    "name": "nginx-deployment",
+    "labels": {
+      "app": "nginx"
+    }
+  },
+  "spec": {
+    "replicas": 3,
+    "selector": {
+      "matchLabels": {
+        "app": "nginx"
+      }
+    },
+    "template": {
+      "metadata": {
+        "labels": {
+          "app": "nginx"
+        }
+      },
+      "spec": {
+        "containers": [
+          {
+            "name": "nginx",
+            "image": "nginx:1.14.2",
+            "ports": [
+              {
+                "containerPort": 80
+              }
+            ]
+          }
+        ]
+      }
+    }
+  }
+}`),
+			expected: []FileType{
+				FileTypeKubernetes,
+				FileTypeJSON,
+			},
+		},
+		{
 			name: "YAML, no reader",
 			path: "file.yaml",
 			r:    nil,