--- a/pkg/detection/testdata/small.file
+++ b/pkg/detection/testdata/small.file
@@ -1,0 +1,3 @@
+{
+  "content": "foo bar baz"
+}