… pkg/detection/testdata/big.file