Currently in some places we use `io.ReadAll` which can slow things down. We can replace this `io.Copy` calls to improve the performance.

Fixes: https://github.com/aquasecurity/defsec/issues/1108, https://github.com/aquasecurity/trivy/issues/2286, https://github.com/aquasecurity/trivy/issues/3212


## Results

Here a `small file` is 30B and `big file` is 10M.


```shell
$ benchstat io.ReadAll=old.txt io.Copy=new.txt
goos: darwin
goarch: amd64
pkg: github.com/aquasecurity/defsec/pkg/detection
cpu: Intel(R) Core(TM) i7-8569U CPU @ 2.80GHz
                   │  io.ReadAll  │               io.Copy               │
                   │    sec/op    │   sec/op     vs base                │
IsType_SmallFile-8   3.704µ ± 27%   3.707µ ± 1%        ~ (p=0.971 n=10)
IsType_BigFile-8     3.875µ ± 32%   1.973µ ± 2%  -49.07% (p=0.000 n=10)
geomean              3.788µ         2.705µ       -28.60%

                   │  io.ReadAll   │               io.Copy                │
                   │     B/op      │     B/op      vs base                │
IsType_SmallFile-8    6.307Ki ± 0%   5.916Ki ± 0%   -6.19% (p=0.000 n=10)
IsType_BigFile-8     21.352Ki ± 0%   9.644Ki ± 0%  -54.83% (p=0.000 n=10)
geomean               11.60Ki        7.553Ki       -34.91%

                   │ io.ReadAll │              io.Copy               │
                   │ allocs/op  │ allocs/op   vs base                │
IsType_SmallFile-8   59.00 ± 0%   60.00 ± 0%   +1.69% (p=0.000 n=10)
IsType_BigFile-8     15.00 ± 0%   10.00 ± 0%  -33.33% (p=0.000 n=10)
geomean              29.75        24.49       -17.66%
```


## Results with much larger files (10GB)

### Old call graph

<img width="881" alt="image" src="https://user-images.githubusercontent.com/1254783/222854615-4ae9e73b-ec2e-4c35-9645-00450e139aa5.png">


### New call graph (observe the hot path no longer includes reading the entire file upfront)

<img width="1012" alt="image" src="https://user-images.githubusercontent.com/1254783/222854580-c5101041-fee6-4394-9442-699870ea2ede.png">


Signed-off-by: Simar <simar@linux.com>