@simar7 2023-03-04
If using this change in Trivy, the results are as follows

## Test scan setup

```shell
tree
.
├── big.file
├── huge.json
├── huge.yaml
└── small.file

➜  testdata git:(big-files-bug) ✗ ls -lrth *.file  
-rw-r--r--  1 simar  staff    30B Mar  2 15:07 small.file
-rw-------  1 simar  staff   5.0K Mar  3 15:20 big.file
➜  testdata git:(big-files-bug) ✗ ls -lrth huge.*    
-rw-------  1 simar  staff   1.0G Mar  3 16:59 huge.yaml
-rw-------  1 simar  staff   1.0G Mar  3 17:08 huge.json

```

## Results

### Old
```shell
time /Users/simar/repos/trivy/trivy.old config --format table --exit-code 1 --severity HIGH,CRITICAL --debug .
2023/03/03 17:08:51 profile: cpu profiling enabled, cpu.pprof
2023-03-03T17:08:51.111-0800    DEBUG   Severities: ["HIGH" "CRITICAL"]
2023-03-03T17:08:51.165-0800    DEBUG   cache dir:  /Users/simar/Library/Caches/trivy
2023-03-03T17:08:51.165-0800    INFO    Misconfiguration scanning is enabled
2023-03-03T17:08:51.166-0800    DEBUG   Policies successfully loaded from disk
2023-03-03T17:08:51.166-0800    DEBUG   Walk the file tree rooted at '.' in parallel
2023-03-03T17:09:16.487-0800    DEBUG   OS is not detected.
2023-03-03T17:09:16.487-0800    INFO    Detected config files: 0
2023/03/03 17:09:16 profile: cpu profiling disabled, cpu.pprof
/Users/simar/repos/trivy/trivy.old config --format table --exit-code 1    .  

12.29s user 13.80s system 98% cpu 26.458 total
```

Old Flamegraph to visualize the time spent in `IsType`, 19.92s.
<img width="715" alt="image" src="https://user-images.githubusercontent.com/1254783/222868535-269c909d-bcfb-46b0-be24-b69812c377a2.png">



### New
```shell
time /Users/simar/repos/trivy/trivy.new config --format table --exit-code 1 --severity HIGH,CRITICAL --debug .
2023/03/03 17:10:30 profile: cpu profiling enabled, cpu.pprof
2023-03-03T17:10:30.606-0800    DEBUG   Severities: ["HIGH" "CRITICAL"]
2023-03-03T17:10:30.657-0800    DEBUG   cache dir:  /Users/simar/Library/Caches/trivy
2023-03-03T17:10:30.658-0800    INFO    Misconfiguration scanning is enabled
2023-03-03T17:10:30.658-0800    DEBUG   Policies successfully loaded from disk
2023-03-03T17:10:30.659-0800    DEBUG   Walk the file tree rooted at '.' in parallel
2023-03-03T17:10:43.655-0800    DEBUG   OS is not detected.
2023-03-03T17:10:43.655-0800    INFO    Detected config files: 0
2023/03/03 17:10:43 profile: cpu profiling disabled, cpu.pprof
/Users/simar/repos/trivy/trivy.new config --format table --exit-code 1    .  

9.03s user 10.74s system 113% cpu 17.372 total

```

With a flamegraph you can see the time spent in `IsType` is quite less as compared to before, 13.26s

<img width="718" alt="image" src="https://user-images.githubusercontent.com/1254783/222868492-4574e3d7-3baa-4ed7-aff0-7e85b56b510c.png">