@simar7 2023-03-04
We can further improve this within Trivy as well. There are places in the code where files can be potentially loaded all at once into memory, causing slowdowns. But for that I'll make a different PR.